Change grid edit functionality
Completely removed the edit modal as it was ugly and one never needs to change that much at once anyway. Instead the view modal got repurposed to also be able to edit fields. The checks from the edit modal were preserved though & overwriting of other grids should still be prevented.