Flatten structure: move all files to root level
Removed the nested structure as it doesn't comply with the production server. Now all the folders are in root.

Also removed some debug files.
Also removed leftover files from the move to php

The config files are now in a separate folder in private. Should make it more secure and enable the development on a local server & production on the university's lamp server. Should also make the compability between the two environments via git much easier & allow direct pushes