Rework grid preparation saving to add a button in database view to include new grids
Initial problem: When saving the form initially, only grids were saved in the database that had the include bool set to true. Had to change this, so that the new button can just flip that boolean from false to true if needed.
Also, the sample logic was wrong. Sample should always be across all grids and no logic is needed to check if that sample is different now & we will thus also not have a problem with session_id key not populating properly anymore.

Fixed the initial saving/updating logic. I hope it is correct now. It now always assumes that this is the same across all grids. If this needs changing in the future, we will need to revisit this logic.

Button added & also triggers reloading of view to make the change visible.