Added sample foreign key to make lookups easier
This new foreign key makes sample information lookups much easier as it doesn't have to go over the grid_preparation table anymore to find the values.

As the endpoint changed, had to update both the sessioncontroller for the input form, the grid modal to edit values and the validation behind it.