Add Microscope Input modal
Added a modal in the Admin tab to enter session details at the microscope. It is using the same modal structure as the other ones but overly wide on purpose. That way it is not constrained to the container - which is better for computer screens and because the table will become quite wide.

So far it only handles screening information, the rest I will add later.