Add microscope session to admin panel for an overview
Meant as an overview of all the sessions, not separated by user.
Style is the same as all the other tables with the foldout.

Also needed to change the style of the foldout to make it longer to avoid cutting off slots on the bottom. The overflow to the right of the container still needs to be handled correctly though.