Auto-populate with last microscope parameters for users
Had to add a new endpoint for this. Also had to make sure, that the autopopulation doesn't trigger each time the edit modal is called and avoid overwriting existing.

To inform the user of the autopopulation, a nice popup with the parameters is called and ask if these parameters should be applied

Currently doesn't check for other user's entries though